+++ src/react-components/src/components/00-base/Container/Container.tsx
@@ -1,0 +1,15 @@
+
+import * as React from "react";
+import bem from "../../../utils/bem";
+
+export interface ContainerProps {
+  modifiers?: string[];
+  blockName?: string;
+}
+
+export default function Container(props: React.PropsWithChildren<ContainerProps>) {
+  const { modifiers = [], blockName } = props;
+  return (
+    <div className={bem("container")}>{props.children}</div>
+  );
+}