+++ src/react-components/src/components/03-organisms/EditionsList/EditionsList.tsx
@@ -1,0 +1,30 @@
+// OL-24 Editions List
+import * as React from "react";
+import bem from "../../../utils/bem";
+import EditionCard, { EditionDetails } from "../../02-molecules/Cards/EditionCard";
+
+export interface SearchResultsListProps {
+  baseClass?: string;
+  modifiers?: string[];
+  blockName?: string;
+  editions: EditionDetails[];
+}
+export default function SearchResultsList(props: SearchResultsListProps) {
+  const { baseClass = "edition-list", modifiers = [], blockName, editions } = props;
+
+  return <ul className={bem(baseClass, modifiers, blockName)}>
+    {editions.map((edition, index) =>
+      <li key={`edition-item-${index}`}>
+        <EditionCard
+          id={`edition-${index}`}
+          blockName={blockName}
+          modifiers={["list"]}
+          coverUrl={edition.coverUrl}
+          editionHeadingElement={edition.editionYearHeading}
+          editionInfo={[edition.publisherAndLocation, edition.language, edition.license]}
+          readOnlineLink={edition.readOnlineLink}
+          downloadLink={edition.downloadLink} />
+      </li>
+    )}
+  </ul>;
+}