+++ src/styles/CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This repo is in Prerelease. When it is released, this project will adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ========
-## Prerelease
+## PRERELEASE 
+### Added 
+- Styles for `editions-list`
+
+## [0.0.8] - 2020-01-15
 ### Added
 - Styles for `edition-card`
 - Styles for `header-with-search`